requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
opencv-python==4.10.0.84
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3